consolidate example files
