+++ examples/directedTwitterAgent/index.ts
@@ -1,16 +1,64 @@
 import { config } from '../../src/config/index.js';
 import { createLogger } from '../../src/utils/logger.js';
 import { validateLocalHash } from '../../src/agents/tools/utils/localHashStorage.js';
-import { orchestratorRunner } from './agent.js';
+import {
+  createOrchestratorRunner,
+  OrchestratorRunner,
+} from '../../src/agents/workflows/orchestrator/orchestratorWorkflow.js';
+import { createTools } from '../../src/agents/workflows/orchestrator/tools.js';
+import { createDirectedTwitterAgentTool } from '../../src/agents/workflows/directedTwitter/directedTwitterAgentTool.js';
+import { createPrompts } from '../../src/agents/workflows/orchestrator/prompts.js';
+import { LLMProvider } from '../../src/services/llm/types.js';
+import { PruningParameters } from '../../src/agents/workflows/orchestrator/types.js';
+import { LLMFactory } from '../../src/services/llm/factory.js';
 import { HumanMessage } from '@langchain/core/messages';
 
-export const logger = createLogger('directed-twitter-agent');
+const logger = createLogger('directed-twitter-agent');
 
-const runner = await orchestratorRunner();
-const initalMessage = `As a social media manager, you are expected to interact with twitter periodically in order to maintain social engagement. Use your judgement how frequently you should run these interactions. Run the twitter workflow to handle twitter related tasks.
-`;
+const orchestatorConfig = async () => {
+  const directedTwitterAgent = createDirectedTwitterAgentTool();
+  const namespace = 'orchestrator';
+  const { tools } = createTools();
+  const prompts = await createPrompts({ selfSchedule: true });
+  const pruningParameters: PruningParameters = {
+    maxWindowSummary: 30,
+    maxQueueSize: 50,
+  };
+  const model = LLMFactory.createModel({
+    provider: LLMProvider.ANTHROPIC,
+    model: 'claude-3-5-sonnet-latest',
+    temperature: 0,
+  });
+  return {
+    model,
+    namespace,
+    tools: [...tools, directedTwitterAgent],
+    prompts,
+    pruningParameters,
+  };
+};
+
+const orchestratorConfig = await orchestatorConfig();
+const orchestratorRunner = (() => {
+  let runnerPromise: Promise<OrchestratorRunner> | undefined = undefined;
+  return async () => {
+    if (!runnerPromise) {
+      runnerPromise = createOrchestratorRunner(
+        orchestratorConfig.model,
+        orchestratorConfig.tools,
+        orchestratorConfig.prompts,
+        orchestratorConfig.namespace,
+        orchestratorConfig.pruningParameters,
+      );
+    }
+    return runnerPromise;
+  };
+})();
 
 const main = async () => {
+  const runner = await orchestratorRunner();
+  const initalMessage = `As a social media manager, you are expected to interact with twitter periodically in order to maintain social engagement. Use your judgement how frequently you should run these interactions. Run the twitter workflow to handle twitter related tasks.`;
+
   try {
     await validateLocalHash();
